requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1; python_version >= '3.7'
appdirs==1.4.4
async-timeout==4.0.3; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
frozenlist==1.4.0; python_version >= '3.8'
idna==3.4; python_version >= '3.5'
maxminddb==2.4.0
multidict==6.0.4; python_version >= '3.7'
pyyaml==6.0.1
requests==2.31.0
urllib3==2.0.4; python_version >= '3.7'
yarl==1.9.2; python_version >= '3.7'
